requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
httpx>=0.26.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
openai>=1.12.0
google-genai>=0.1.0
anthropic>=0.18.0
python-dotenv>=1.0.1
